move replicate calls to api file
